+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Support for `debug` widget
 * Global state object exposed
 * `All` options in `select-checkboxes`
+* Better default label in `select-checkboxes`
 
 ### Changed